separate the utility into a separate file
