+++ packages/vintex/src/pager/index.ts
@@ -1,1 +1,2 @@
 export * from './pager';
+export * from './utils';
