+++ packages/vintex/src/pager/pager.tsx
@@ -4,32 +4,11 @@ import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 import { type TreeNode } from '../tree';
 import { getLocaleUrl, type I18nConfig } from '../utils';
-
-type Route = Required<Pick<TreeNode, 'title' | 'path'>>;
+import { getPaths } from './utils';
 
 interface PagerProps extends I18nConfig {
   sidebar: TreeNode[];
 }
-
-/** Returns all the paths that are in a sidebar */
-export const getPaths = (sidebar: TreeNode[]): Route[] => {
-  const paths: Route[] = [];
-
-  const obtenerPaths = (arr: TreeNode[]) => {
-    arr.forEach(item => {
-      if (item.path) {
-        paths.push({ title: item.title, path: item.path });
-      }
-
-      if (item.children) {
-        obtenerPaths(item.children);
-      }
-    });
-  };
-
-  obtenerPaths(sidebar);
-  return paths;
-};
 
 export const Pager = ({ sidebar, lang, defaultLang }: PagerProps) => {
   const pathname = usePathname();
